Use different components for layers and region overlays
