+++ web/src/components/map/ActiveLayers.vue
@@ -2,17 +2,12 @@
 import { ref } from "vue";
 import draggable from "vuedraggable";
 import {
-  selectedRegions,
-  regionGroupingType,
   availableDerivedRegions,
-  cancelRegionGrouping,
-  regionGroupingActive,
   activeMapLayerIds,
   currentCity,
   currentMapDataSource,
 } from "@/store";
 import { getDataSourceFromLayerId, updateVisibleLayers } from "@/layers";
-import { postDerivedRegion, listDerivedRegions } from "@/api/rest";
 import { watch } from "vue";
 import { MapDataSource, MapDataSourceArgs } from "@/data";
 
@@ -24,6 +19,12 @@ watch(activeMapLayerIds, (val, oldVal) => {
     layerMenuActive.value = true;
   }
 });
+
+function clearActiveLayers() {
+  activeMapLayerIds.value = [];
+  updateVisibleLayers();
+  layerMenuActive.value = false;
+}
 
 function getLayerName(layerId: string) {
   const dataSource = getDataSourceFromLayerId(layerId);
@@ -64,33 +65,9 @@ function setCurrentMapDataSource(layerId: string) {
 
   currentMapDataSource.value = new MapDataSource(args);
 }
-
-// Region Controls
-const newRegionName = ref("");
-async function createDerivedRegion() {
-  if (selectedRegions.value.length === 0) {
-    throw new Error("Cannot created derived region with no selected regions");
-  }
-  if (regionGroupingType.value === null) {
-    throw new Error("Region grouping type is null");
-  }
-
-  const city = selectedRegions.value[0].city;
-  await postDerivedRegion(
-    newRegionName.value,
-    city,
-    selectedRegions.value.map((reg) => reg.id),
-    regionGroupingType.value
-  );
-
-  // Close dialog
-  cancelRegionGrouping();
-  availableDerivedRegions.value = await listDerivedRegions();
-}
 </script>
 
 <template>
-  <!-- Active layers -->
   <v-menu
     v-model="layerMenuActive"
     persistent
@@ -104,7 +81,10 @@ async function createDerivedRegion() {
       </v-btn>
     </template>
     <v-card rounded="lg" class="mt-2">
-      <v-card-title>Active Layers</v-card-title>
+      <v-card-title>
+        Active Layers
+        <v-icon @click="clearActiveLayers">mdi-playlist-remove</v-icon>
+      </v-card-title>
       <draggable
         v-model="activeMapLayerIds"
         @change="updateVisibleLayers"
@@ -128,48 +108,4 @@ async function createDerivedRegion() {
       </draggable>
     </v-card>
   </v-menu>
-
-  <template v-if="regionGroupingActive">
-    <v-divider />
-    <!-- Region grouping -->
-    <v-card v-if="regionGroupingActive">
-      <v-card-title class="text-capitalize pb-0">
-        <v-icon size="small">mdi-vector-{{ regionGroupingType }}</v-icon>
-        Performing {{ regionGroupingType }} Grouping
-      </v-card-title>
-      <v-card-subtitle>
-        Grouping {{ selectedRegions.length }} Regions
-      </v-card-subtitle>
-
-      <v-row no-gutters class="px-2 mt-2">
-        <v-text-field
-          v-model="newRegionName"
-          hide-details
-          label="New Region Name"
-        />
-      </v-row>
-      <v-card-actions>
-        <v-row no-gutters>
-          <v-spacer />
-          <v-btn
-            variant="tonal"
-            color="error"
-            prepend-icon="mdi-cancel"
-            @click="cancelRegionGrouping"
-          >
-            Cancel
-          </v-btn>
-          <v-btn
-            color="success"
-            variant="flat"
-            prepend-icon="mdi-check"
-            :disabled="selectedRegions.length < 2 || !newRegionName"
-            @click="createDerivedRegion"
-          >
-            Save
-          </v-btn>
-        </v-row>
-      </v-card-actions>
-    </v-card>
-  </template>
 </template>
